Test stats on the "dst" after it have been synced
The discussion in the issue #5 revealed the problems in mtime sync.

The first was is that test code tries to set mtime on all source files
to an hour ago, but messes it up by doing changing contents of the
directories after their mtimes have been set. It also misses setting
mtime for the "src" directory itself.

These changes led to the discovery of the actual bug in the test code:
it tests mtime for the "dst" directory before it has been synced. I
removed that untimely test, added a new call to sync "dst" directory,
and added stats checks for it after it.